pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "seo-agent"
version = "1.0.0"
description = "Web Analysis Tool for SEO, AEO, and GEO"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "lxml>=5.1.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "types-beautifulsoup4>=4.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
